requirements.txt
Django==3.0.8
djangorestframework==3.11.0
django-cors-headers==3.4.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2==2.8.5
whitenoise==5.2.0